--- gatsby/src/components/loadable.tsx
@@ -1,3 +1,0 @@
-import Loadable from "@loadable/component"
-
-export const loadableP5 = Loadable(() => import("react-p5-wrapper"))
